requirements.txt
alembic==1.0.10
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
coverage==4.5.3
entrypoints==0.3
filelock==3.0.12
flake8==3.7.7
Flask==1.0.3
Flask-API==1.1
Flask-JWT-Extended==3.18.2
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
importlib-metadata==0.17
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
Mako==1.0.10
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
packaging==19.0
pep8==1.7.1
pluggy==0.12.0
psycopg2-binary==2.8.2
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.0
pytest==4.6.0
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.3
toml==0.10.0
tox==3.12.1
typed-ast==1.3.5
virtualenv==16.6.0
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.1
zipp==0.5.1
